Structure: Restructure settings
