+++ src/components/settings/LogoutButton/index.ts
@@ -1,0 +1,1 @@
+export * from './LogoutButton.component';
