+++ src/components/settings/SettingsPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingsPanel.component';
