+++ src/components/settings/ThemeSwitch/ThemeSwitch.component.tsx
