+++ src/components/settings/ThemeSwitch/index.ts
@@ -1,0 +1,1 @@
+export * from './ThemeSwitch.component';
